module-task1: * edit task interface
