+++ src/common/interfaces/task/task.interface.ts
@@ -10,8 +10,6 @@ interface ITask {
   color: TaskColor;
   isArchive: boolean;
   isFavorite: boolean;
-  isDueDate?: boolean;
-  isRepeating?: boolean;
 }
 
 export { ITask };
