+++ src/view/task-edit/common.ts
@@ -1,7 +1,7 @@
+import { getRandomId } from '~/helpers';
+import { ITask } from '~/common/interfaces';
 import { TaskColor } from '~/common/enums';
 import { TASK_DEFAULT_REPEATING } from '~/common/constants';
-import { ITask } from '~/common/interfaces';
-import { getRandomId } from '~/helpers';
 
 const EMPTY_TASK: ITask = {
   id: getRandomId(),
@@ -16,3 +16,10 @@ const EMPTY_TASK: ITask = {
 };
 
 export { EMPTY_TASK };
+
+interface IRawTask extends ITask {
+  isDueDate?: boolean;
+  isRepeating?: boolean;
+}
+
+export { IRawTask };
